pyproject.toml
[project]
name = "mini-books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2pdf>=0.1.8",
    "opencv-python>=4.11.0.86",
    "python-docx>=1.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.11",
]
